+++ base/task/thread_pool/job_task_source.h
@@ -96,9 +96,10 @@ class BASE_EXPORT JobTaskSource : public TaskSource {
   // ever modified under a lock or read atomically (optimistic read).
   class State {
    public:
-    static constexpr size_t kCanceledMask = 1;
-    static constexpr size_t kWorkerCountBitOffset = 1;
-    static constexpr size_t kWorkerCountIncrement = 1 << kWorkerCountBitOffset;
+    static constexpr uint32_t kCanceledMask = 1;
+    static constexpr int kWorkerCountBitOffset = 1;
+    static constexpr uint32_t kWorkerCountIncrement = 1
+                                                      << kWorkerCountBitOffset;
 
     struct Value {
       size_t worker_count() const { return value >> kWorkerCountBitOffset; }
